+++ src/smc_mallows_new_users_funs.cpp
@@ -38,13 +38,14 @@ void smc_mallows_new_users_augment_partial(
 
       // fill in missing ranks based on choice of augmentation method
       Rcpp::List proposal{};
-      if (aug_method == "random") {
+      const bool pseudo = is_pseudo(aug_method, metric);
+      if (!pseudo) {
         // create new augmented ranking by sampling remaining ranks from set uniformly
         partial_ranking.elem(find_nonfinite(partial_ranking)) = shuffle(missing_ranks);
 
         aug_rankings(span(jj), span::all, span(ii)) = partial_ranking;
         aug_prob(ii) = divide_by_fact(aug_prob(ii), missing_ranks.n_elem);
-      } else if ((aug_method == "pseudolikelihood") && ((metric == "footrule") || (metric == "spearman"))) {
+      } else {
         // randomly permute the unranked items to give the order in which they will be allocated
         uvec item_ordering = shuffle(unranked_items);
         const rowvec& rho_s = rho_samples(span(ii), span::all, span(tt + 1));
@@ -57,8 +58,6 @@ void smc_mallows_new_users_augment_partial(
         const double& f_prob = proposal["forward_prob"];
         aug_rankings(span(jj), span::all, span(ii)) = a_rank;
         aug_prob(ii) = aug_prob(ii) * f_prob;
-      } else {
-        Rcpp::stop("Combined choice of metric and aug_method is incompatible");
       }
     }
   }