+++ NEWS.md
@@ -1,6 +1,7 @@
 # BayesMallows (development versions)
 
 * Replaced deprecated ggplot2::aes_ function with ggplot2::aes.
+* Refactoring of SMC functions (#257)
 
 # BayesMallows 1.2.2