+++ NEWS.md
@@ -1,4 +1,4 @@
-# BayesMallows (development version)
+# BayesMallows 1.3.0
 
 * Added heat_plot() function (#255)
 * Replaced deprecated ggplot2::aes_ function with ggplot2::aes.